debugging

+++ main.c
@@ -22,8 +22,8 @@ int main(int argc, char** argv){
 	// Parse command line arguments
 	int N, num_threads, block_size, tile_size;
 	parse_args(argc, argv, &N, &num_threads, &block_size, &tile_size);
-	size_t n_blocks=N/block_size;
-	size_t block_area=block_size*block_size;
+	int n_blocks=N/block_size;
+	int block_area=block_size*block_size;
 
 	// Initialize MPI
 	int required = MPI_THREAD_FUNNELED, provided;
